+++ app/server/server.js
@@ -66,14 +66,15 @@ var run = function () {
 
   // check environment
   var port = process.env.PORT ? parseInt(process.env.PORT) : 80;
-  var mongo_url = process.env.MONGO_URL;
-  if (!mongo_url)
+
+  // check for a valid MongoDB URL right away
+  if (!process.env.MONGO_URL)
     throw new Error("MONGO_URL must be set in environment");
 
   // webserver
   var app = connect.createServer();
   var static_cacheable_path = path.join(bundle_dir, 'static_cacheable');
-  if (path.existsSync(static_cacheable_path))
+  if (fs.existsSync(static_cacheable_path))
     app.use(gzippo.staticGzip(static_cacheable_path, {clientMaxAge: 1000 * 60 * 60 * 24 * 365}));
   app.use(gzippo.staticGzip(path.join(bundle_dir, 'static')));
 
@@ -87,9 +88,6 @@ var run = function () {
   __meteor_runtime_config__ = {};
   Fiber(function () {
     // (put in a fiber to let Meteor.db operations happen during loading)
-
-    // pass in database info
-    __meteor_bootstrap__.mongo_url = mongo_url;
 
     // load app code
     _.each(info.load, function (filename) {