+++ packages/mongo-livedata/mongo_livedata_tests.js
@@ -14,7 +14,7 @@ testAsyncMulti("mongo-livedata - database failure reporting", [
     };
 
     _.each(["insert", "remove", "update"], function (op) {
-      if (Meteor.is_server) {
+      if (Meteor.isServer) {
         test.throws(function () {
           ftc[op]({fail: true});
         });
@@ -22,7 +22,7 @@ testAsyncMulti("mongo-livedata - database failure reporting", [
         ftc[op]({fail: true}, expect(exception));
       }
 
-      if (Meteor.is_client) {
+      if (Meteor.isClient) {
         ftc[op]({fail: true}, expect(exception));
 
         // This would log to console in normal operation.
@@ -37,7 +37,7 @@ testAsyncMulti("mongo-livedata - database failure reporting", [
 Tinytest.addAsync("mongo-livedata - basics", function (test, onComplete) {
   var run = test.runId();
   var coll, coll2;
-  if (Meteor.is_client) {
+  if (Meteor.isClient) {
     coll = new Meteor.Collection(null); // local, unmanaged
     coll2 = new Meteor.Collection(null); // local, unmanaged
   } else {
@@ -62,7 +62,7 @@ Tinytest.addAsync("mongo-livedata - basics", function (test, onComplete) {
   });
 
   var captureObserve = function (f) {
-    if (Meteor.is_client) {
+    if (Meteor.isClient) {
       f();
     } else {
       var fence = new Meteor._WriteFence;
@@ -106,10 +106,28 @@ Tinytest.addAsync("mongo-livedata - basics", function (test, onComplete) {
   test.equal(coll.findOne({run: run}, {sort: {x: -1}, skip: 0}).x, 4);
   test.equal(coll.findOne({run: run}, {sort: {x: -1}, skip: 1}).x, 1);
 
+  // sleep function from fibers docs.
+  var sleep = function(ms) {
+    var fiber = Fiber.current;
+    setTimeout(function() {
+      fiber.run();
+    }, ms);
+    Fiber.yield();
+  };
+
   var cur = coll.find({run: run}, {sort: ["x"]});
   var total = 0;
   cur.forEach(function (doc) {
     total *= 10;
+    if (Meteor.isServer) {
+      // Verify that the callbacks from forEach run sequentially and that
+      // forEach waits for them to complete (issue# 321). If they do not run
+      // sequentially, then the second callback could execute during the first
+      // callback's sleep sleep and the *= 10 will occur before the += 1, then
+      // total (at test.equal time) will be 5. If forEach does not wait for the
+      // callbacks to complete, then total (at test.equal time) will be 0.
+      sleep(5);
+    }
     total += doc.x;
     // verify the meteor environment is set up here
     coll2.insert({total:total});
@@ -155,7 +173,7 @@ Tinytest.addAsync("mongo-livedata - fuzz test", function(test, onComplete) {
 
   var run = test.runId();
   var coll;
-  if (Meteor.is_client) {
+  if (Meteor.isClient) {
     coll = new Meteor.Collection(null); // local, unmanaged
   } else {
     coll = new Meteor.Collection("livedata_test_collection_"+run);
@@ -201,7 +219,7 @@ Tinytest.addAsync("mongo-livedata - fuzz test", function(test, onComplete) {
   };
 
   var finishObserve = function (f) {
-    if (Meteor.is_client) {
+    if (Meteor.isClient) {
       f();
     } else {
       var fence = new Meteor._WriteFence;
@@ -220,7 +238,7 @@ Tinytest.addAsync("mongo-livedata - fuzz test", function(test, onComplete) {
     var max_counters = _.clone(counters);
 
     finishObserve(function () {
-      if (Meteor.is_server)
+      if (Meteor.isServer)
         obs._suspendPolling();
 
       // Do a batch of 1-10 operations
@@ -253,7 +271,7 @@ Tinytest.addAsync("mongo-livedata - fuzz test", function(test, onComplete) {
           max_counters.remove++;
         }
       }
-      if (Meteor.is_server)
+      if (Meteor.isServer)
         obs._resumePolling();
 
     });
@@ -269,7 +287,7 @@ Tinytest.addAsync("mongo-livedata - fuzz test", function(test, onComplete) {
       test.isTrue(max_counters[k] >= counters[k], k);
     });
 
-    Tinytest.defer(doStep);
+    Meteor.defer(doStep);
   };
 
   doStep();
@@ -279,14 +297,14 @@ Tinytest.addAsync("mongo-livedata - fuzz test", function(test, onComplete) {
 Tinytest.addAsync("mongo-livedata - scribbling", function (test, onComplete) {
   var run = test.runId();
   var coll;
-  if (Meteor.is_client) {
+  if (Meteor.isClient) {
     coll = new Meteor.Collection(null); // local, unmanaged
   } else {
     coll = new Meteor.Collection("livedata_test_collection_"+run);
   }
 
   var runInFence = function (f) {
-    if (Meteor.is_client) {
+    if (Meteor.isClient) {
       f();
     } else {
       var fence = new Meteor._WriteFence;