+++ packages/sass/sass_tests.js
@@ -2,14 +2,13 @@
 Tinytest.add("sass - presence", function(test) {
 
   var d = OnscreenDiv(Meteor.render(function() {
-    return '<p class="sass-unlucky-left-border"></p>'; }));
+    return '<p class="sass-dashy-left-border"></p>'; }));
   d.node().style.display = 'block';
 
   var p = d.node().firstChild;
-  var leftBorder = getStyleProperty(p, 'border-left-width');
-  test.equal(leftBorder, "13px");
+  var leftBorder = getStyleProperty(p, 'border-left-style');
+  test.equal(leftBorder, "dashed");
 
   d.kill();
 
 });
-