+++ packages/spark/package.js
@@ -7,12 +7,6 @@ Package.on_use(function (api) {
   api.use(['underscore', 'uuid', 'domutils', 'liverange', 'universal-events'],
           'client');
 
-  // XXX Depends on jquery because we need a selector engine to resolve
-  // event maps. What would be nice is, if you've included jquery or
-  // zepto, use one of those; if not, ship our own copy of sizzle (but,
-  // you still want the event object normalization that jquery provides?)
-  api.use('jquery');
-
   api.add_files(['spark.js', 'patch.js', 'convenience.js'], 'client');
 });