+++ packages/templating/deftemplate.js
@@ -17,10 +17,11 @@
       return Spark.list(
         arg,
         function (item) {
-          return Spark.labelBranch(item._id || null, function () {
-            var html = Spark.isolate(_.bind(options.fn, null, item));
-            return Spark.setDataContext(item, html);
-          });
+          return Spark.labelBranch(
+            item._id || Spark.UNIQUE_LABEL, function () {
+              var html = Spark.isolate(_.bind(options.fn, null, item));
+              return Spark.setDataContext(item, html);
+            });
         },
         function () {
           return options.inverse ?