+++ packages/tinytest/tinytest.js
@@ -38,7 +38,7 @@ _.extend(TestCaseResults.prototype, {
     var self = this;
 
     if (self.stop_at_offset === 0) {
-      if (Meteor.is_client) {
+      if (Meteor.isClient) {
         // Only supported on the browser for now..
         var now = (+new Date);
         debugger;
@@ -248,7 +248,7 @@ _.extend(TestCase.prototype, {
                                           onException(e);
                                       }, stop_at_offset);
 
-    Tinytest.defer(function () {
+    Meteor.defer(function () {
       try {
         if (self.async) {
           self.func(results, function () {
@@ -351,7 +351,7 @@ _.extend(TestRun.prototype, {
     var testGroups = _.values(
       _.groupBy(self.manager.ordered_tests,
                 function(t) {
-                  if (Meteor.is_server)
+                  if (Meteor.isServer)
                     return "SERVER";
                   if (t.async)
                     return "ASYNC";
@@ -416,40 +416,7 @@ globals.Tinytest = {
 
   addAsync: function (name, func) {
     TestManager.addCase(new TestCase(name, func, true));
-  },
-
-  // Tinytest has versions of setTimeout et al. that preserve the
-  // environment with write fence on the server.  Unlike how
-  // Meteor.setTimeout etc. are currently implemented, we *DO*
-  // want to block method satisfaction.
-
-  setTimeout: function (f, duration) {
-    return setTimeout(Meteor.bindEnvironment(f, function (e) {
-      Meteor._debug("Exception from setTimeout callback:", e.stack);
-    }), duration);
-  },
-
-  setInterval: function (f, duration) {
-    return setInterval(Meteor.bindEnvironment(f, function (e) {
-      Meteor._debug("Exception from setInterval callback:", e);
-    }), duration);
-  },
-
-  clearInterval: function(x) {
-    return clearInterval(x);
-  },
-
-  clearTimeout: function(x) {
-    return clearTimeout(x);
-  },
-
-  defer: function (f) {
-    // Older Firefox will pass an argument to the setTimeout callback
-    // function, indicating the "actual lateness." It's non-standard,
-    // so for defer, standardize on not having it.
-    globals.Tinytest.setTimeout(function () {f();}, 0);
   }
-
 };
 
 // Run every test, asynchronously. Runs the test in the current