+++ History.md
@@ -1,30 +1,90 @@
 
 ## vNEXT
 
+## v0.4.1
+
+* New `email` smart package, with [`Email.send`](http://docs.meteor.com/#email)
+  API.
+
+* Upgrade Node from 0.6.17 to 0.8.8, as well as many Node modules in the dev
+  bundle; those that are user-exposed are:
+  * coffee-script: 1.3.3 (from 1.3.1)
+  * stylus: 0.29.0 (from 0.28.1)
+  * nib: 0.8.2 (from 0.7.0)
+
+* All publicly documented APIs now use `camelCase` rather than
+  `under_scores`. The old spellings continue to work for now. New names are:
+  - `Meteor.isClient`/`isServer`
+  - `this.isSimulation` inside a method invocation
+  - `Meteor.deps.Context.onInvalidate`
+  - `Meteor.status().retryCount`/`retryTime`
+
+* Spark improvements
+  * Optimize selector matching for event maps.
+  * Fix `Spark._currentRenderer` behavior in timer callbacks.
+  * Fix bug caused by interaction between `Template.foo.preserve` and
+    `{{#constant}}`. #323
+  * Allow `{{#each}}` over a collection of objects without `_id`. #281
+  * Spark now supports Firefox 3.6.
+  * Added a script to build a standalone spark.js that does not depend on
+    Meteor (it depends on jQuery or Sizzle if you need IE7 support,
+    and otherwise is fully standalone).
+
+* Database writes from within `Meteor.setTimeout`/`setInterval`/`defer` will be
+  batched with other writes from the current method invocation if they start
+  before the method completes.
+
+* Make `Meteor.Cursor.forEach` fully synchronous even if the user's callback
+  yields. #321.
+
+* Recover from exceptions thrown in `Meteor.publish` handlers.
+
+* Upgrade bootstrap to version 2.1.1. #336, #337, #288, #293
+
+* Change the implementation of the `meteor deploy` password prompt to not crash
+  Emacs M-x shell.
+
+* Optimize `LocalCollection.remove(id)` to be O(1) rather than O(n).
+
+* Optimize client-side database performance when receiving updated data from the
+  server outside of method calls.
+
+* Better error reporting when a package in `.meteor/packages` does not exist.
+
+* Better error reporting for coffeescript. #331
+
+* Better error handling in `Handlebars.Exception`.
+
+
+Patches contributed by GitHub users fivethirty, tmeasday, and xenolf.
+
+
 ## v0.4.0
 
 * Merge Spark, a new live page update engine
   * Breaking API changes
-    * Input elements no longer preserved based on `id` and `name` attributes. Use [`preserve`](http://docs.meteor.com/#template_preserve) instead.
-
-    * All `Meteor.ui` functions removed. Use `Meteor.render`, `Meteor.renderList`, and [Spark](https://github.com/meteor/meteor/wiki/Spark) functions instead.
-
-    * New template functions (eg. `created`, `rendered`, etc) may collide with existing helpers. Use `Template.foo.helpers()` to avoid conflicts.
-
-    * New syntax for declaring event maps. Use `Template.foo.events({...})`. For backwards compatibility, both syntaxes are allowed for now.
-
+     * Input elements no longer preserved based on `id` and `name`
+       attributes. Use [`preserve`](http://docs.meteor.com/#template_preserve)
+       instead.
+     * All `Meteor.ui` functions removed. Use `Meteor.render`,
+       `Meteor.renderList`, and
+       [Spark](https://github.com/meteor/meteor/wiki/Spark) functions instead.
+     * New template functions (eg. `created`, `rendered`, etc) may collide with
+       existing helpers. Use `Template.foo.helpers()` to avoid conflicts.
+     * New syntax for declaring event maps. Use
+       `Template.foo.events({...})`. For backwards compatibility, both syntaxes
+       are allowed for now.
   * New Template features
-
-    * Allow embedding non-Meteor widgets (eg. Google Maps) using [`{{#constant}}`](http://docs.meteor.com/#constant)
-
-    * Callbacks when templates are rendered. See http://docs.meteor.com/#template_rendered
-
-    * Explicit control of which nodes are preserved during re-rendering. See http://docs.meteor.com/#template_preserve
-
-    * Easily find nodes within a template in event handlers and callbacks. See http://docs.meteor.com/#template_find
-
-    * Allow parts of a template to be independently reactive with the [`{{#isolate}}`](http://docs.meteor.com/#isolate) block helper.
-
+     * Allow embedding non-Meteor widgets (eg. Google Maps) using
+       [`{{#constant}}`](http://docs.meteor.com/#constant)
+     * Callbacks when templates are rendered. See
+       http://docs.meteor.com/#template_rendered
+     * Explicit control of which nodes are preserved during re-rendering. See
+       http://docs.meteor.com/#template_preserve
+     * Easily find nodes within a template in event handlers and callbacks. See
+       http://docs.meteor.com/#template_find
+     * Allow parts of a template to be independently reactive with the
+       [`{{#isolate}}`](http://docs.meteor.com/#isolate) block helper.
 
 * Use PACKAGE_DIRS environment variable to override package location. #227
 
@@ -58,7 +118,8 @@
   * Allow functions in helper arguments.
   * Change helper nesting rules to allow functions as arguments.
   * Fix `{{this.foo}}` to never invoke helper `foo`.
-  * Make event handler `this` reflect the node that matched the selector instead of the event target node.
+  * Make event handler `this` reflect the node that matched the selector instead
+    of the event target node.
   * Fix keyword arguments to helpers.
 
 * Add `nib` support to stylus package. #175
@@ -86,7 +147,8 @@
 
 * Form control improvements
   * Fix reactive radio buttons in Internet Explorer.
-  * Fix reactive textareas to update consistently across browsers, matching text field behavior.
+  * Fix reactive textareas to update consistently across browsers, matching text
+    field behavior.
 
 * `http` package bug fixes:
   * Send correct Content-Type when POSTing `params` from the server. #172
@@ -96,18 +158,24 @@
 
 * Fix intermittent "Cursor is closed" mongo error.
 
-* Fix "Cannot read property 'nextSibling' of null" error in certain nested templates. #142
+* Fix "Cannot read property 'nextSibling' of null" error in certain nested
+  templates. #142
 
-* Add heartbeat timer on the client to notice when the server silently goes away.
+* Add heartbeat timer on the client to notice when the server silently goes
+  away.
 
 
 ## v0.3.6
 
-* Rewrite event handling. `this` in event handlers now refers to the data context of the element that generated the event, *not* the top-level data context of the template where the event is declared.
+* Rewrite event handling. `this` in event handlers now refers to the data
+  context of the element that generated the event, *not* the top-level data
+  context of the template where the event is declared.
 
-* Add /websocket endpoint for raw websockets. Pass websockets through development mode proxy.
+* Add /websocket endpoint for raw websockets. Pass websockets through
+  development mode proxy.
 
-* Simplified API for Meteor.connect, which now receives a URL to a Meteor app rather than to a sockjs endpoint.
+* Simplified API for Meteor.connect, which now receives a URL to a Meteor app
+  rather than to a sockjs endpoint.
 
 * Fix livedata to support subscriptions with overlapping documents.
 
@@ -131,7 +199,8 @@
 
 * Make `backbone` package also run on the server.
 
-* Add `bare` option to coffee-script compilation so variables can be shared between multiple coffee-script file. #85
+* Add `bare` option to coffee-script compilation so variables can be shared
+  between multiple coffee-script file. #85
 
 * Upgrade many dependency versions. User visible highlights:
  * node.js 0.6.15
@@ -156,7 +225,8 @@
 
 * Fix issue with spaces in directory names. #39
 
-* Workaround browser caching issues in development mode by using query parameters on all JavaScript and CSS requests.
+* Workaround browser caching issues in development mode by using query
+  parameters on all JavaScript and CSS requests.
 
 * Many documentation and test fixups.